Struct bonsaidb_client::Client

pub struct Client { /* private fields */ }

Client for connecting to a BonsaiDb server.

Connecting via QUIC

The URL scheme to connect via QUIC is bonsaidb. If no port is specified, port 5645 is assumed.

With a valid TLS certificate

let client = Client::build(Url::parse("bonsaidb://my-server.com")?).finish()?;

With a Self-Signed Pinned Certificate

When using install_self_signed_certificate(), clients will need the contents of the pinned-certificate.der file within the database. It can be specified when building the client:

let certificate =
    Certificate::from_der(std::fs::read("mydb.bonsaidb/pinned-certificate.der")?)?;
let client = Client::build(Url::parse("bonsaidb://localhost")?)
    .with_certificate(certificate)
    .finish()?;

Connecting via WebSockets

WebSockets are built atop the HTTP protocol. There are two URL schemes for WebSockets:

ws: Insecure WebSockets. Port 80 is assumed if no port is specified.

wss: Secure WebSockets. Port 443 is assumed if no port is specified.

Without TLS

let client = Client::build(Url::parse("ws://localhost")?).finish()?;

With TLS

let client = Client::build(Url::parse("wss://my-server.com")?).finish()?;

Using a Api

Our user guide has a section on creating and using an Api.

// `bonsaidb_core` is re-exported to `bonsaidb::core` or `bonsaidb_client::core`.
use bonsaidb_core::{
    api::{Api, Infallible},
    schema::{ApiName, Qualified},
};
use serde::{Deserialize, Serialize};

#[derive(Serialize, Deserialize, Debug)]
pub struct Ping;

#[derive(Serialize, Deserialize, Clone, Debug)]
pub struct Pong;

impl Api for Ping {
    type Response = Pong;
    type Error = Infallible;

    fn name() -> ApiName {
        ApiName::private("ping")
    }
}

let client = Client::build(Url::parse("bonsaidb://localhost")?).finish()?;
let Pong = client.send_api_request_async(&Ping).await?;

Receiving out-of-band messages from the server

If the server sends a message that isn’t in response to a request, the client will invoke it’s api callback:

let client = Client::build(Url::parse("bonsaidb://localhost")?)
    .with_api_callback(ApiCallback::<Ping>::new(|result: Pong| async move {
        println!("Received out-of-band Pong");
    }))
    .finish()?;

Implementations

impl Client

pub fn build(url: Url) -> Builder

Returns a builder for a new client connecting to url.

impl Client

pub fn new(url: Url) -> Result<Self, Error>

Initialize a client connecting to url. This client can be shared by cloning it. All requests are done asynchronously over the same connection.

If the client has an error connecting, the first request made will present that error. If the client disconnects while processing requests, all requests being processed will exit and return Error::Disconnected. The client will automatically try reconnecting.

The goal of this design of this reconnection strategy is to make it easier to build resilliant apps. By allowing existing Client instances to recover and reconnect, each component of the apps built can adopt a “retry-to-recover” design, or “abort-and-fail” depending on how critical the database is to operation.

pub async fn send_api_request_async<Api: Api>(
&self,
request: &Api
) -> Result<Api::Response, ApiError<Api::Error>>

Sends an api request.

pub fn invoke_api_request<Api: Api>(&self, request: &Api) -> Result<(), Error>

Sends an api request without waiting for a result. The response from the server will be ignored.

pub fn send_api_request<Api: Api>(
&self,
request: &Api
) -> Result<Api::Response, ApiError<Api::Error>>

Sends an api request.

pub async fn effective_permissions(&self) -> Option<Permissions>

Returns the current effective permissions for the client. Returns None if unauthenticated.